Removing the "RiakVectorClock" annotation: too dangerous to keep around.
Basically, annotating an object with @RiakVectorClock and keeping it around
(in a cache, or wherever) then saving it again will create siblings, if
other writes have occurred.

It's a lot safer to simply NEVER annotate objects with @RiakVectorClock,
and let users keep track of the vector clock (if they want to) through
the Response and StoreRequest objects.